perf(canvas): widen Crossing.dir to avoid a store-forwarding stall
The insertion sort in sortCrossings copies whole Crossing structs; with
dir: i8 that 4-byte load overlapped the 1-byte store just made, and the
failed store-to-load forwarding cost ~15 cycles per scanline. Fast thick
lines had dropped to 0.6x of their pre-#407 speed; they are back at
parity (vertical w=3: 2.5 -> 1.8 us).

+++ src/canvas/Canvas.zig
@@ -1595,10 +1595,11 @@ pub fn Canvas(comptime T: type) type {
             }
         };
 
-        /// Where an edge crosses a scanline, with its winding direction.
+        /// Where an edge crosses a scanline, with its winding direction. `dir` is a full
+        /// word so that sorting's whole-struct copies never read a byte-wide store.
         const Crossing = struct {
             x: f32,
-            dir: i8,
+            dir: i32,
 
             fn lessThan(_: void, a: Crossing, b: Crossing) bool {
                 return a.x < b.x;
